Keep the memory plugin type import in the baselined memory-v2.ts; memory-validate.ts only registers
