+++ assistant/src/cli/commands/memory/memory-v2.ts
@@ -12,9 +12,9 @@
  *     from the current skill set.
  *   - `activation` — refresh persisted activation state for every
  *     conversation that has a stored row.
- *   - `validate`: the same read-only report as the top-level
- *     `memory validate` (`memory-validate.ts`), registered here too while
- *     this namespace exists.
+ *   - `validate`: {@link runMemoryValidate}, also registered as the top-level
+ *     `memory validate` (`memory-validate.ts`); kept here while this
+ *     namespace exists.
  */
 
 import type { Command } from "commander";
@@ -26,6 +26,7 @@ import type {
   MemoryV2EmaScoresResult,
   MemoryV2ReembedSkillsResult,
   MemoryV2SimulateRouterResult,
+  MemoryV2ValidateResult,
 } from "../../../plugins/defaults/memory/src/memory-v2-routes.js";
 import type { ComparisonReport } from "../../../plugins/defaults/memory/v2/harness/runner.js";
 import { subcommand } from "../../lib/cli-command-help.js";
@@ -34,7 +35,6 @@ import {
   renderComparisonReport,
   renderTurnTrace,
 } from "./memory-v2-compare-render.js";
-import { runMemoryValidate } from "./memory-validate.js";
 
 // ---------------------------------------------------------------------------
 // Helpers
@@ -58,6 +58,60 @@ async function runBackfillOp(op: MemoryV2BackfillOp): Promise<void> {
   }
 
   log.info(`Queued ${op} job: ${result.result!.jobId}`);
+}
+
+/**
+ * Print the validate report; sets a non-zero exit code on any violation.
+ * Registered as `memory validate` (`memory-validate.ts`) and `memory v2
+ * validate`.
+ */
+export async function runMemoryValidate(): Promise<void> {
+  const result = await cliIpcCall<MemoryV2ValidateResult>(
+    "memory_v2_validate",
+    { body: {} },
+  );
+
+  if (!result.ok) {
+    log.error(result.error ?? "Failed to validate memory state");
+    process.exitCode = 1;
+    return;
+  }
+
+  const report = result.result!;
+  log.info(`Pages: ${report.pageCount}`);
+  log.info(`Edges: ${report.edgeCount}`);
+  log.info(
+    `Dangling links: ${
+      report.danglingLinks.length === 0 ? "none" : report.danglingLinks.length
+    }`,
+  );
+  for (const d of report.danglingLinks) {
+    log.info(`  - ${d.from} → ${d.to} (${d.kind})`);
+  }
+  log.info(
+    `Oversized pages: ${
+      report.oversizedPages.length === 0 ? "none" : report.oversizedPages.length
+    }`,
+  );
+  for (const p of report.oversizedPages) {
+    log.info(`  - ${p.slug}: ${p.chars} chars`);
+  }
+  log.info(
+    `Parse failures: ${
+      report.parseFailures.length === 0 ? "none" : report.parseFailures.length
+    }`,
+  );
+  for (const p of report.parseFailures) {
+    log.info(`  - ${p.slug}: ${p.error}`);
+  }
+
+  if (
+    report.danglingLinks.length > 0 ||
+    report.oversizedPages.length > 0 ||
+    report.parseFailures.length > 0
+  ) {
+    process.exitCode = 1;
+  }
 }
 
 // ---------------------------------------------------------------------------
